Refactored Preference navigation tests for compact and non-compact mode
